Don't use environment variables troughout the documentation
The idea with this is that no example or test in the
documentation or function tests (Those use mockito), so
there isn't really a need to have it there.

Instead what we will do is have the tests that perform
real requests on a separate tests folder and don't use mockito
on those.